+++ src/apps/todo-list/store.ts
@@ -221,24 +221,6 @@ export async function changeTodoItemDone(todo: TodoItem, done: boolean): Promise
   )
 }
 
-export function getDoneInputClass(todoItem: TodoItem) {
-  let cls = 'checkbox checkbox-sm'
-  switch (todoItem.priority) {
-    case '高':
-      cls += ' checkbox-error'
-      break
-    case '中':
-      cls += ' checkbox-warning'
-      break
-    case '低':
-      cls += ' checkbox-info'
-      break
-    default:
-      break
-  }
-  return cls
-}
-
 /**
  * 删除一个待办事项
  *
